Create a color palette from an image

Derive qualitative, sequential and divergent color palettes from an image on disk or at a URL.

image_pal(file, n = 9, type = c("qual", "seq", "div"), bw = c(0, 1),
  brightness = c(0, 1), saturation = c(0, 1), seq_by = "hsv",
  seed = NULL, plot = FALSE, labels = TRUE, label_size = 1,
  keep_asp = TRUE)

Arguments

file
character, file path or URL to an image.

n
integer, number of colors.

type
character, type of palette: qualitative, sequential or divergent ("qual", "seq", or "div").

bw
a numeric vector of length two giving the lower and upper quantiles to trim trim near-black and near-white colors in RGB space.

brightness
as above, trim possible colors based on brightness in HSV space.

saturation
as above, trim possible colors based on saturation in HSV space.

seq_by
character, sort sequential palette by HSV dimensions in a specific order, e.g., "hsv", "svh". See details.

seed
numeric, set the seed for reproducible results.

plot
logical, plot the palette.

labels
logical, show hex color values in plot.

label_size
numeric, label size in plot.

keep_asp
logical, adjust rectangles in plot to use the image aspect ratio.

Value

character vector of hex colors, optionally draws a plot

Details

Ordering colors is a challenging problem. There are many ways to do it; none are perfect. Color is a multi-dimensional property; any reduction to a a one dimensional color spectrum necessarily removes information.

Creating a sequential palette from an arbitrary image that contains several hues, at different saturation and brightness levels, and making a palette that looks sequential is particularly problematic. This function does a decent job of creating qualitative, sequential and divergent palettes from images, but additional tweaking of function arguments is needed on a case by case basis. This can include trimming the extreme values of the color distribution in terms of brightness, saturation and presence of near-black/white colors as pre-processing steps. There is also variation in possible palettes from a given image, depending on the image complexity and other properties, though you can set the random seed for reproducibility.

Trimming color distribution

Some pre-processing can be done to limit undesirable colors from ending up in a palette. bw specifically drops near-black and near-white colors as soon as the image is loaded by looking at the average values in RGB space. brightness and saturation trimming are applied subsequently to trim lower and upper quantiles of the HSV value and saturation, respectively. If you have already trimmed black and white, keep in mind these two arguments will trim further from what remains of the color distribution.

Choosing appropriate palette type

Keep in mind that many images simple do not make sense to try to derive sensible color palettes from. For images that do lend themselves to a useful color palette derivation, some may only make sense to consider for a divergent palette, or an increasing/decreasing sequential palette, or only a qualitative palette if there are too many colors that are difficult to order. For divergent palettes in particular, it is recommended to trim white, e.g. bw = c(0, 0.9), depending on the white space of a given image, since the divergent palettes are centered on white.

Sorting sequential palettes

seq_by = "hsv" orders the final palette by hue, then saturation, then value (brightness). This default is not meant to be ideal for all images. It work better in cases where sequential palettes may contain several distinct hues, but not much variation in saturation or brightness. However, for example, palettes derived from an image with relatively little variation in hue may appear more sorted to the human eye if ordered by hue last using "svh" or "vsh", depending on whether you want the palette to appear to transition more from lower saturation or lower brightness to the predominant hue.

Examples

# Only run in interactive R session
if(interactive()){
  set.seed(1)
  x <- paste0(system.file(package = "imgpalr"), "/",
    c("blue-yellow", "purples", "colors"), ".jpg")

  # Three palette types, one image
  # Focus on bright, saturated colors for divergent palette:
  image_pal(x[1], type = "div",
    saturation = c(0.75, 1), brightness = c(0.75, 1), plot = TRUE)

  # Remove colors too close to black and white for qualitative palette:
  image_pal(x[1], type = "qual", bw = c(0.25, 0.9), plot = TRUE)

   # A challenging sequential mapping
  image_pal(x[1], type = "seq", saturation = c(0.2, 1),
    brightness = c(0.5, 1), seq_by = "hsv", plot = TRUE)

  # An image with one main hue;
  # sequential is most sensible, but not necessarily best sorted by hue
  image_pal(x[2], type = "seq", seq_by = "hsv", plot = TRUE)
  image_pal(x[2], type = "seq", seq_by = "svh", plot = TRUE)
  image_pal(x[2], type = "seq", seq_by = "vsh", plot = TRUE)

  # An image with one several hues;
  # divergent is not sensible here, sequential should be by hue
  image_pal(x[3], type = "qual", brightness = c(0.4, 1), plot = TRUE)
  image_pal(x[3], type = "seq", bw = c(0.2, 1), saturation = c(0.2, 1), plot = TRUE)
}